# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: EduManage School SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /schools/{schoolId}/students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List students in a school
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
        - name: classId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      summary: Add a student
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentBody"
      responses:
        "201":
          description: Student created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /schools/{schoolId}/students/{studentId}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get student details
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    put:
      operationId: updateStudent
      tags: [students]
      summary: Update a student
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStudentBody"
      responses:
        "200":
          description: Updated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Delete a student
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # CLASSES
  /schools/{schoolId}/classes:
    get:
      operationId: listClasses
      tags: [classes]
      summary: List classes
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Class"
    post:
      operationId: createClass
      tags: [classes]
      summary: Create a class
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClassBody"
      responses:
        "201":
          description: Class created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
  /schools/{schoolId}/classes/{classId}:
    delete:
      operationId: deleteClass
      tags: [classes]
      summary: Delete a class
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
        - name: classId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ATTENDANCE
...
components:
  schemas:
...
    Student:
      type: object
      properties:
        id:
          type: integer
        schoolId:
          type: integer
        name:
          type: string
        studentNumber:
          type: string
        classId:
          type: integer
          nullable: true
        className:
          type: string
          nullable: true
        category:
          type: string
          enum: [regular, bus, scholarship, staff_child]
        gender:
          type: string
          enum: [male, female, other]
          nullable: true
        dateOfBirth:
          type: string
          format: date
          nullable: true
        parentName:
          type: string
          nullable: true
        parentPhone:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: string
          format: date-time
        totalFeeExpected:
          type: number
          nullable: true
        totalFeePaid:
          type: number
          nullable: true
        arrears:
          type: number
          nullable: true
      required: [id, schoolId, name, studentNumber, category, status, createdAt]
    CreateStudentBody:
      type: object
      properties:
        name:
          type: string
        studentNumber:
          type: string
        classId:
          type: integer
        category:
          type: string
          enum: [regular, bus, scholarship, staff_child]
        gender:
          type: string
          enum: [male, female, other]
        dateOfBirth:
          type: string
          format: date
        parentName:
          type: string
        parentPhone:
          type: string
      required: [name, studentNumber, category]
    UpdateStudentBody:
      type: object
      properties:
        name:
          type: string
        classId:
          type: integer
        category:
          type: string
          enum: [regular, bus, scholarship, staff_child]
        gender:
          type: string
          enum: [male, female, other]
        dateOfBirth:
          type: string
          format: date
        parentName:
          type: string
        parentPhone:
          type: string
        status:
          type: string
          enum: [active, inactive]
    Class:
      type: object
      properties:
        id:
          type: integer
        schoolId:
          type: integer
        name:
          type: string
        grade:
          type: string
          nullable: true
        teacherId:
          type: integer
          nullable: true
        teacherName:
          type: string
          nullable: true
        studentCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, schoolId, name, studentCount, createdAt]
    CreateClassBody:
      type: object
      properties:
        name:
          type: string
        grade:
          type: string
        teacherId:
          type: integer
      required: [name]
...